+++ Assets/Scripts/Stats/EnemyStats.cs
@@ -13,7 +13,7 @@ public class EnemyStats : CharacterStats
         enemy = GetComponent<Enemy>();
     }
 
-    public override void TakeDamage(int _damage)
+    public override void TakeDamage(float _damage)
     {
         base.TakeDamage(_damage);
     }
